optimize deploy node process
